Squelch some additional warnings

+++ include/sir/helpers.h
@@ -439,8 +439,8 @@ bool _sir_getchar(char* input);
  */
 # define _sir_snprintf_trunc(dst, size, ...) \
     do { \
-      volatile size_t _n = size; \
-      if (!snprintf(dst, _n, __VA_ARGS__)) { (void)_n; }; \
+      volatile size_t _n = (size_t)size; \
+      if (!snprintf(dst, (int)(size_t)_n, __VA_ARGS__)) { (void)_n; }; \
     } while (false)
 
 /**
